🎨Design: SearchBar의 뒤로가기 버튼 분리(#16) SearchBar 컴포넌트의 뒤로 가기 버튼을 재사용 가능성이 높다고 판단되어 BackButton이라는 이름의 컴포넌트로 분리 Related: #16
