+++ src/components/search/SearchBar.tsx
@@ -1,20 +1,16 @@
 import styled from 'styled-components';
 
-import arrowBack from '../../assets/icon/icon-arrow.svg';
+import { BackButton } from 'components/common/BackButton';
 import search from '../../assets/icon/icon-search.svg';
 
 export const SearchBar: React.FC = () => {
-  const handleBackButtonClick = (): void => {
-    console.log('BackButton has been clicked!');
-  };
-
   const handleSearchButtonClick = (): void => {
     console.log('SearchButton has been clicked!');
   };
 
   return (
     <SearchBarContainer>
-      <BackButton onClick={handleBackButtonClick} />
+      <BackButton />
       <SearchArea>
         <SearchInput type="text" placeholder="검색어를 입력하세요" />
         <SearchButton onClick={handleSearchButtonClick} />
@@ -28,17 +24,8 @@ const SearchBarContainer = styled.div`
   margin: 0 auto 20px auto;
   width: 90%;
   align-items: center;
-`;
 
-const BackButton = styled.button`
-  background-image: url(${arrowBack});
-  background-repeat: no-repeat;
-  background-position: center;
-  background-size: contain;
-  width: 20px;
-  height: 20px;
-  cursor: pointer;
-  margin-right: 10px;
+  padding-top: 20px;
 `;
 
 const SearchArea = styled.div`
